# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/kyc/requirements": {
      "get": {
        "operationId": "KycController_requirements_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Pièces acceptées selon mon pays et contraintes de fichier (US-3.1)",
        "tags": ["KYC"]
      }
    },
...
    "/api/v1/kyc/submit": {
      "post": {
        "operationId": "KycController_submit[0]_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Soumettre pièce d’identité (recto/verso) + selfie caméra (US-3.1)",
        "tags": ["KYC"]
      }
    },
...
    "/api/v1/kyc/tier3": {
      "post": {
        "operationId": "KycController_tier3_v1",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Demander le niveau 3 (justificatif de domicile, source de revenus, selfie) — revue manuelle",
        "tags": ["KYC"]
      }
    },
    "/api/v1/kyc/me": {
      "get": {
        "operationId": "KycController_me[0]_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Mon statut KYC et l’historique de mes soumissions",
        "tags": ["KYC"]
      }
    },
...
    "/api/v1/kyc/reviews": {
      "get": {
        "operationId": "KycController_queue[0]_v1",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Dossiers à revoir, triés par ancienneté, avec SLA",
        "tags": ["KYC"]
      }
    },
...
    "/api/v1/kyc/requests/{id}": {
      "get": {
        "operationId": "KycController_detail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Détail d’un dossier (documents, scores, alertes, historique)",
        "tags": ["KYC"]
      }
    },
    "/api/v1/kyc/documents/{id}": {
      "get": {
        "operationId": "KycController_document_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Image d’un document déchiffrée (accès journalisé)",
        "tags": ["KYC"]
      }
    },
    "/api/v1/kyc/requests/{id}/decision": {
      "post": {
        "operationId": "KycController_decide_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "enum": ["APPROVE"]
                      },
                      "annotation": {
                        "type": "string",
                        "minLength": 10,
                        "maxLength": 2000
                      }
                    },
                    "required": ["action", "annotation"],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "enum": ["REJECT"]
                      },
                      "category": {
                        "type": "string",
                        "enum": [
                          "DOCUMENT_ILLISIBLE",
                          "DOCUMENT_EXPIRE",
                          "DOCUMENT_FALSIFIE",
                          "FACE_MATCH_ECHOUE",
                          "DOUBLON_CONFIRME",
                          "AML_MATCH_CONFIRME",
                          "INFORMATION_INCOHERENTE",
                          "TYPE_DOCUMENT_NON_ACCEPTE",
                          "AUTRE"
                        ]
                      },
                      "comment": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 2000
                      }
                    },
                    "required": ["action", "category", "comment"],
                    "additionalProperties": false
                  },
                  {
                    "type": "object",
                    "properties": {
                      "action": {
                        "type": "string",
                        "enum": ["REQUEST_SUPPLEMENT"]
                      },
                      "comment": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 2000
                      }
                    },
                    "required": ["action", "comment"],
                    "additionalProperties": false
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Décision : accepter (annotation ≥ 10), rejeter (catégorie + commentaire), compléments",
        "tags": ["KYC"]
      }
    },
    "/api/v1/kyc/duplicates": {
      "get": {
        "operationId": "KycController_duplicates_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Alertes de doublons biométriques (US-3.4)",
        "tags": ["KYC"]
      }
    },
    "/api/v1/kyc/duplicates/{id}/resolve": {
      "post": {
        "operationId": "KycController_resolveDuplicate_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "resolution": {
                    "type": "string",
                    "enum": ["CONFIRMED", "DISMISSED"]
                  },
                  "comment": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  }
                },
                "required": ["resolution", "comment"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["KYC"]
      }
    },
    "/api/v1/kyc/aml-matches": {
      "get": {
        "operationId": "KycController_aml_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Correspondances AML / PEP / sanctions (US-3.5)",
        "tags": ["KYC"]
      }
    },
    "/api/v1/kyc/aml-matches/{id}/resolve": {
      "post": {
        "operationId": "KycController_resolveAml_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "resolution": {
                    "type": "string",
                    "enum": ["CONFIRMED_MATCH", "FALSE_POSITIVE"]
                  },
                  "comment": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  }
                },
                "required": ["resolution", "comment"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": ["KYC"]
      }
    },
    "/api/v1/kyc/verify": {
      "post": {
        "operationId": "KycController_verify_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "annotation": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 2000
                  }
                },
                "required": ["requestId", "annotation"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Accepter un dossier (alias de requests/{id}/decision, APPROVE)",
        "tags": ["KYC"]
      }
    },
    "/api/v1/kyc/reject": {
      "post": {
        "operationId": "KycController_reject_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "requestId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "category": {
                    "type": "string",
                    "enum": [
                      "DOCUMENT_ILLISIBLE",
                      "DOCUMENT_EXPIRE",
                      "DOCUMENT_FALSIFIE",
                      "FACE_MATCH_ECHOUE",
                      "DOUBLON_CONFIRME",
                      "AML_MATCH_CONFIRME",
                      "INFORMATION_INCOHERENTE",
                      "TYPE_DOCUMENT_NON_ACCEPTE",
                      "AUTRE"
                    ]
                  },
                  "comment": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 2000
                  }
                },
                "required": ["requestId", "category", "comment"],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Rejeter un dossier (alias de requests/{id}/decision, REJECT)",
        "tags": ["KYC"]
      }
    },
    "/api/v1/kyc/{memberId}": {
      "get": {
        "operationId": "KycController_ofMember_v1",
        "parameters": [
          {
            "name": "memberId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Statut et historique KYC d’un membre (personnel, accès journalisé)",
        "tags": ["KYC"]
      }
    },
...
  },
  "info": {
    "title": "TontineMoney API",
    "description": "API REST versionnée du MVP TontineMoney (monolithe modulaire). Erreurs au format Problem Details (RFC 9457). Montants en chaînes décimales (unités majeures) + devise ISO 4217. En-tête Idempotency-Key obligatoire sur les POST financiers.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tm_rt"
      }
    },
...
  }
}